+++ lib/segment/src/vector_storage/multi_dense/appendable_mmap_multi_dense_vector_storage.rs
@@ -27,7 +27,7 @@ use crate::vector_storage::dense::appendable_dense_vector_storage::{
     open_appendable_memmap_vector_storage_byte, open_appendable_memmap_vector_storage_full,
     open_appendable_memmap_vector_storage_half,
 };
-use crate::vector_storage::turbo::multi::open_appendable_turbo_multi_vector_storage;
+use crate::vector_storage::turbo::multi_turbo::open_appendable_turbo_multi_vector_storage;
 use crate::vector_storage::turbo::open_appendable_turbo_vector_storage;
 use crate::vector_storage::{
     MultiVectorStorage, MultiVectorStorageRead, VectorOffsetType, VectorStorage, VectorStorageEnum,
@@ -470,7 +470,7 @@ pub fn open_appendable_memmap_vector_storage(
         ),
         VectorStorageDatatype::Turbo4 => {
             open_appendable_turbo_vector_storage(vector_storage_path, size, distance, populate)
-                .map(|s| VectorStorageEnum::DenseTurbo(Box::new(s)))
+                .map(|s| VectorStorageEnum::DenseTurboAppendableMemmap(Box::new(s)))
         }
     }
 }
