+++ lib/segment/src/vector_storage/query_scorer/turbo_multi_query_scorer.rs
@@ -6,7 +6,7 @@ use quantization::turboquant::EncodedQueryTQ;
 
 use crate::data_types::vectors::MultiDenseVectorInternal;
 use crate::vector_storage::query_scorer::QueryScorer;
-use crate::vector_storage::turbo::multi::TurboMultiScoring;
+use crate::vector_storage::TurboMultiScoring;
 
 /// Asymmetric MaxSim raw scorer for a multivector TurboQuant storage
 /// ([`TurboMultiScoring`]).
