+++ lib/segment/src/vector_storage/query_scorer/turbo_query_scorer.rs
@@ -5,7 +5,7 @@ use quantization::turboquant::EncodedQueryTQ;
 
 use crate::data_types::vectors::DenseVector;
 use crate::vector_storage::query_scorer::QueryScorer;
-use crate::vector_storage::turbo::TurboScoring;
+use crate::vector_storage::TurboScoring;
 
 /// Asymmetric raw scorer for a dense TurboQuant storage ([`TurboScoring`]).
 ///
