+++ lib/segment/src/vector_storage/tests/mod.rs
@@ -3,5 +3,7 @@ mod async_raw_scorer;
 mod custom_query_scorer_equivalency;
 mod test_appendable_dense_vector_storage;
 mod test_appendable_multi_dense_vector_storage;
+mod test_appendable_multi_turbo_vector_storage;
 mod test_appendable_sparse_vector_storage;
+mod test_appendable_turbo_vector_storage;
 mod utils;
