+++ lib/segment/src/vector_storage/turbo/multi_turbo/read_only/live_reload.rs
@@ -3,14 +3,14 @@ use common::sorted_slice::SortedSlice;
 use common::types::PointOffsetType;
 use common::universal_io::UniversalRead;
 
-use super::ReadOnlyChunkedTurboVectorStorage;
+use super::ReadOnlyChunkedMultiTurboVectorStorage;
 use crate::common::live_reload::LiveReload;
 use crate::common::operation_error::OperationResult;
 
-impl<S: UniversalRead> LiveReload for ReadOnlyChunkedTurboVectorStorage<S> {
+impl<S: UniversalRead> LiveReload for ReadOnlyChunkedMultiTurboVectorStorage<S> {
     type Fs = S::Fs;
 
-    /// Reload the chunked vectors, apply `deleted_points`, and fold in the
+    /// Reload the vectors and offsets, apply `deleted_points`, and fold in the
     /// persisted deletion of each appended offset — a live point may have a
     /// deleted vector slot recorded only on disk.
     fn live_reload(
@@ -21,6 +21,8 @@ impl<S: UniversalRead> LiveReload for ReadOnlyChunkedTurboVectorStorage<S> {
         hw_counter: &HardwareCounterCell,
     ) -> OperationResult<()> {
         self.storage
+            .live_reload(fs, deleted_points, new_points, hw_counter)?;
+        self.offsets
             .live_reload(fs, deleted_points, new_points, hw_counter)?;
         self.deleted.insert_all(deleted_points);
         self.deleted.reload_appended::<S>(fs, new_points)?;
