+++ lib/segment/src/vector_storage/turbo/read_only/immutable/live_reload.rs
@@ -1,0 +1,25 @@
+use common::counter::hardware_counter::HardwareCounterCell;
+use common::sorted_slice::SortedSlice;
+use common::types::PointOffsetType;
+use common::universal_io::UniversalRead;
+
+use super::ReadOnlyImmutableTurboVectorStorage;
+use crate::common::live_reload::LiveReload;
+use crate::common::operation_error::OperationResult;
+
+impl<S: UniversalRead> LiveReload for ReadOnlyImmutableTurboVectorStorage<S> {
+    type Fs = S::Fs;
+
+    /// Vector data is immutable, so only the in-memory deletion flags are patched
+    /// from the authoritative `deleted_points`; `fs` and `new_points` are unused.
+    fn live_reload(
+        &mut self,
+        _fs: &S::Fs,
+        deleted_points: &SortedSlice<'_, PointOffsetType>,
+        _new_points: &SortedSlice<'_, PointOffsetType>,
+        _hw_counter: &HardwareCounterCell,
+    ) -> OperationResult<()> {
+        self.deleted.insert_all(deleted_points);
+        Ok(())
+    }
+}
