+++ lib/segment/src/vector_storage/turbo/read_only/immutable/read_ops.rs
@@ -7,16 +7,16 @@ use common::universal_io::{UniversalRead, UserData};
 use quantization::EncodedStorage;
 use quantization::turboquant::EncodedQueryTQ;
 
-use super::super::shared;
-use super::ReadOnlyChunkedTurboVectorStorage;
+use super::ReadOnlyImmutableTurboVectorStorage;
 use crate::common::operation_error::OperationResult;
 use crate::data_types::named_vectors::CowVector;
 use crate::data_types::vectors::DenseVector;
 use crate::types::{Distance, VectorStorageDatatype};
+use crate::vector_storage::turbo::shared;
 use crate::vector_storage::vector_storage_base::VectorStorageRead;
 use crate::vector_storage::{DenseTQVectorStorageRead, TurboScoring};
 
-impl<S: UniversalRead> VectorStorageRead for ReadOnlyChunkedTurboVectorStorage<S> {
+impl<S: UniversalRead> VectorStorageRead for ReadOnlyImmutableTurboVectorStorage<S> {
     fn size_of_available_vectors_in_bytes(&self) -> usize {
         self.available_vector_count() * self.quantized_vector_size()
     }
@@ -48,10 +48,12 @@ impl<S: UniversalRead> VectorStorageRead for ReadOnlyChunkedTurboVectorStorage<S
     ) {
         let (user_data, point_offsets): (Vec<U>, Vec<PointOffsetType>) = keys.into_iter().unzip();
 
-        self.storage.for_each_batch(&point_offsets, |idx, bytes| {
-            let vector = shared::dequantize_vector(&self.quantizer, self.dim, &bytes);
-            callback(user_data[idx], point_offsets[idx], vector);
-        });
+        self.storage
+            .for_each_in_batch(&point_offsets, |idx, bytes| {
+                let vector = shared::dequantize_vector(&self.quantizer, self.dim, bytes);
+                callback(user_data[idx], point_offsets[idx], vector);
+            })
+            .expect("read TQ vectors");
     }
 
     fn get_vector_opt<P: AccessPattern>(&self, key: PointOffsetType) -> Option<CowVector<'_>> {
@@ -80,7 +82,7 @@ impl<S: UniversalRead> VectorStorageRead for ReadOnlyChunkedTurboVectorStorage<S
     }
 }
 
-impl<S: UniversalRead> DenseTQVectorStorageRead for ReadOnlyChunkedTurboVectorStorage<S> {
+impl<S: UniversalRead> DenseTQVectorStorageRead for ReadOnlyImmutableTurboVectorStorage<S> {
     fn vector_dim(&self) -> usize {
         self.dim
     }
@@ -96,10 +98,9 @@ impl<S: UniversalRead> DenseTQVectorStorageRead for ReadOnlyChunkedTurboVectorSt
     fn for_each_in_dense_tq_batch<F: FnMut(usize, &[u8])>(
         &self,
         keys: &[PointOffsetType],
-        mut f: F,
+        f: F,
     ) -> OperationResult<()> {
-        self.storage.for_each_batch(keys, |idx, bytes| f(idx, &bytes));
-        Ok(())
+        self.storage.for_each_in_batch(keys, f)
     }
 
     fn read_dense_tq_bytes<P: AccessPattern, U: Copy + UserData>(
@@ -109,10 +110,9 @@ impl<S: UniversalRead> DenseTQVectorStorageRead for ReadOnlyChunkedTurboVectorSt
     ) -> OperationResult<()> {
         let (user_data, point_offsets): (Vec<U>, Vec<PointOffsetType>) = keys.into_iter().unzip();
 
-        self.storage.for_each_batch(&point_offsets, |idx, bytes| {
+        self.storage.for_each_in_batch(&point_offsets, |idx, bytes| {
             callback(user_data[idx], point_offsets[idx], bytes.to_vec());
-        });
-        Ok(())
+        })
     }
 
     fn get_dense_for_requantization(
@@ -129,7 +129,7 @@ impl<S: UniversalRead> DenseTQVectorStorageRead for ReadOnlyChunkedTurboVectorSt
     }
 }
 
-impl<S: UniversalRead> TurboScoring for ReadOnlyChunkedTurboVectorStorage<S> {
+impl<S: UniversalRead> TurboScoring for ReadOnlyImmutableTurboVectorStorage<S> {
     fn preprocess_query(&self, query: DenseVector) -> EncodedQueryTQ {
         shared::preprocess_query(&self.quantizer, self.distance, query)
     }
