+++ lib/segment/src/vector_storage/turbo/read_only/mod.rs
@@ -1,4 +1,4 @@
-//! Read-only counterpart of [`TurboVectorStorage`](super::TurboVectorStorage).
+//! Read-only counterparts of the writable dense TurboQuant storages.
 //!
 //! Opens the TurboQuant-encoded blob and deletion flags over an arbitrary
 //! [`UniversalRead`] backend (mmap / cache / remote), so a read-only segment can
@@ -6,109 +6,46 @@
 //! determined by `(dim, distance)` (fixed TQDT constants), so nothing beyond the
 //! encoded bytes and flags is read from disk.
 //!
-//! Scoring is reused verbatim from the writable storage via the shared
-//! [`TurboScoring`](super::TurboScoring) trait — this storage only provides the
-//! read surface it needs.
+//! Mirroring the reference read-only dense storages, the on-disk layout is a
+//! *type*, not a runtime flag — the writable layout the segment produced picks
+//! the read-only type at open time:
+//! - [`ReadOnlyImmutableTurboVectorStorage`] — single-file, read-only counterpart
+//!   of [`TurboVectorStorageImpl`](super::TurboVectorStorageImpl), living in the
+//!   nested [`immutable`] submodule,
+//! - [`ReadOnlyChunkedTurboVectorStorage`] — chunked, read-only counterpart of
+//!   [`AppendableMmapTurboVectorStorage`](super::AppendableMmapTurboVectorStorage).
 //!
-//! The implementation is split across submodules the same way the other
+//! Scoring is reused verbatim from the writable storages via the shared
+//! [`TurboScoring`](crate::vector_storage::TurboScoring) trait; the pure
+//! codec/scoring logic lives in [`super::shared`](super::shared).
+//!
+//! The chunked implementation is split across submodules the same way the other
 //! read-only vector storages are:
 //! - [`lifecycle`]: `preopen` / `open` construction,
 //! - [`read_ops`]: retrieval and scoring trait impls,
 //! - [`live_reload`]: picking up a writer's appends and deletions.
 
-use std::borrow::Cow;
-
-use common::types::PointOffsetType;
 use common::universal_io::UniversalRead;
 use quantization::EncodedStorage;
 use quantization::turboquant::quantization::TurboQuantizer;
 
 use crate::common::flags::in_memory_bitvec_flags::InMemoryBitvecFlags;
-use crate::common::operation_error::OperationResult;
-use crate::types::{Distance, MultiVectorConfig};
-use crate::vector_storage::chunked_vectors::ChunkedVectorsRead;
-use crate::vector_storage::multi_dense::appendable_mmap_multi_dense_vector_storage::MultivectorMmapOffset;
+use crate::types::Distance;
 use crate::vector_storage::quantized::quantized_chunked_mmap_storage::QuantizedChunkedStorageRead;
-use crate::vector_storage::quantized::quantized_storage::QuantizedStorage;
 
+mod immutable;
 mod lifecycle;
 mod live_reload;
 mod read_ops;
 
-/// Read-only TurboQuant encoded backend over a [`UniversalRead`] `S`.
-///
-/// Read-only counterpart of the writable
-/// [`TurboEncodedVectorStorage`](super::turbo_encoded_vectors::TurboEncodedVectorStorage):
-/// the storage layout is picked at open time from the segment's storage type.
-enum ReadOnlyTurboEncoded<S: UniversalRead> {
-    /// Single-file layout, written by the non-appendable `Mmap` / `InRamMmap`.
-    Single(QuantizedStorage<S>),
-    /// Chunked layout, written by the appendable `ChunkedMmap` / `InRamChunkedMmap`.
-    Chunked(QuantizedChunkedStorageRead<S>),
-}
+pub use immutable::ReadOnlyImmutableTurboVectorStorage;
 
-impl<S: UniversalRead> ReadOnlyTurboEncoded<S> {
-    fn get_quantized_vector(&self, key: PointOffsetType) -> Cow<'_, [u8]> {
-        match self {
-            Self::Single(s) => s.get_vector_data(key),
-            Self::Chunked(s) => s.get_vector_data(key),
-        }
-    }
-
-    fn get_quantized_vector_opt(&self, key: PointOffsetType) -> Option<Cow<'_, [u8]>> {
-        match self {
-            Self::Single(s) => s.get_vector_data_opt(key),
-            Self::Chunked(s) => s.get_vector_data_opt(key),
-        }
-    }
-
-    /// Run `f` for each vector in the batch, batching the underlying reads
-    /// where the backend supports it (the single-file backend pipelines
-    /// io_uring / prefetches mmap; the chunked backend batches through
-    /// [`EncodedStorage::for_each_batch`]).
-    fn for_each_in_batch<F: FnMut(usize, &[u8])>(
-        &self,
-        keys: &[PointOffsetType],
-        mut f: F,
-    ) -> OperationResult<()> {
-        match self {
-            Self::Single(s) => s.for_each_in_batch(keys, f),
-            Self::Chunked(s) => {
-                s.for_each_batch(keys, |idx, bytes| f(idx, &bytes));
-                Ok(())
-            }
-        }
-    }
-
-    fn vectors_count(&self) -> usize {
-        match self {
-            Self::Single(s) => s.vectors_count(),
-            Self::Chunked(s) => s.vectors_count(),
-        }
-    }
-
-    fn is_on_disk(&self) -> bool {
-        match self {
-            Self::Single(s) => s.is_on_disk(),
-            Self::Chunked(s) => s.is_on_disk(),
-        }
-    }
-
-    fn populate(&self) -> OperationResult<()> {
-        match self {
-            Self::Single(s) => {
-                s.populate();
-                Ok(())
-            }
-            Self::Chunked(s) => s.populate(),
-        }
-    }
-}
-
-/// Read-only TurboQuant dense vector storage over a [`UniversalRead`] backend.
-pub struct ReadOnlyTurboVectorStorage<S: UniversalRead> {
-    /// Read-only encoded vector blob over one of the two backends.
-    storage: ReadOnlyTurboEncoded<S>,
+/// Read-only chunked TurboQuant dense vector storage over a [`UniversalRead`]
+/// backend `S`. Read-only counterpart of the writable
+/// [`AppendableMmapTurboVectorStorage`](super::AppendableMmapTurboVectorStorage).
+pub struct ReadOnlyChunkedTurboVectorStorage<S: UniversalRead> {
+    /// Read-only chunked encoded vector blob.
+    storage: QuantizedChunkedStorageRead<S>,
     /// Quantizer used to de/quantize and score; rebuilt from `(dim, distance)`.
     quantizer: TurboQuantizer,
     /// Persisted soft-deletion flags, materialized in memory.
@@ -119,45 +56,12 @@ pub struct ReadOnlyTurboVectorStorage<S: UniversalRead> {
     dim: usize,
 }
 
-impl<S: UniversalRead> std::fmt::Debug for ReadOnlyTurboVectorStorage<S> {
+impl<S: UniversalRead> std::fmt::Debug for ReadOnlyChunkedTurboVectorStorage<S> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ReadOnlyTurboVectorStorage")
+        f.debug_struct("ReadOnlyChunkedTurboVectorStorage")
             .field("dim", &self.dim)
             .field("distance", &self.distance)
             .field("total_vector_count", &self.storage.vectors_count())
-            .field("deleted_count", &self.deleted.count())
-            .finish_non_exhaustive()
-    }
-}
-
-/// Read-only TurboQuant multivector vector storage over a [`UniversalRead`]
-/// backend. Read-only counterpart of
-/// [`TurboMultiVectorStorage`](super::multi::TurboMultiVectorStorage).
-///
-/// Multivectors are always stored in the appendable chunked layout, so — unlike
-/// the dense storage — there is a single backend.
-pub struct ReadOnlyTurboMultiVectorStorage<S: UniversalRead> {
-    /// Flat inner-vector space: one fixed-size encoded record per inner vector.
-    storage: QuantizedChunkedStorageRead<S>,
-    /// Maps each point to its record range in the inner space.
-    offsets: ChunkedVectorsRead<MultivectorMmapOffset, S>,
-    /// Quantizer used to de/quantize and score; rebuilt from `(dim, distance)`.
-    quantizer: TurboQuantizer,
-    /// Persisted soft-deletion flags, materialized in memory.
-    deleted: InMemoryBitvecFlags,
-    /// Distance used for scoring / query preprocessing.
-    distance: Distance,
-    /// Original (un-padded) inner vector dimensionality.
-    dim: usize,
-    multi_vector_config: MultiVectorConfig,
-}
-
-impl<S: UniversalRead> std::fmt::Debug for ReadOnlyTurboMultiVectorStorage<S> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ReadOnlyTurboMultiVectorStorage")
-            .field("dim", &self.dim)
-            .field("distance", &self.distance)
-            .field("total_vector_count", &self.offsets.len())
             .field("deleted_count", &self.deleted.count())
             .finish_non_exhaustive()
     }
